+++ CHANGELOG.md
@@ -3,6 +3,7 @@ walkers 5.2
 
 - Add Traditional Chinese (by dirtTW)
 - fix special shape access
+- add charged creeper variant
 
 walkers 5.1
 ================